Distributed Parallel Mining of Frequent Patterns for Stream Data
We propose an efficient technique to discover the complete set of recent
frequent patterns from a high-speed data stream. We develop DPFP-stream,
an effective FP-tree-based model using the concept of tilted-time
windows and MapReduce for mining frequent patterns from data streams. We
implement the algorithm on the Storm platform which accepts data streams
emitted by Kafka. Moreover,the in-memory frequent pattern tree with
titled-time windows is stored in Redis which is suitable for large
sequences. Extensive experimental analyses show that our algorithm is
highly efficient in terms of time complexity when finding recent
frequent patterns from a high-speed data stream.